refactor: consolidate compiler types
